Restructure node code
